[Pdm2-4] Refactor authentication logic into service
